+++ lib/wechat/callback.rb
@@ -1,7 +1,11 @@
 require 'wechat/callback/version'
+
 require 'wechat/callback/message_signature'
 require 'wechat/callback/signature'
 require 'wechat/callback/xml_document'
+
+require 'wechat/callback/message_decryption'
+require 'wechat/callback/message_encryption'
 
 module Wechat
   module Callback