+++ CHANGELOG.md
@@ -1,1 +1,10 @@
 # Wechat Callback Change Log 微信回调库变更日志
+
+## v0.1
+1. Signature class
+2. Message Signature class
+3. XML Document class
+
+## v0.2
+1. Message Encryption class
+2. Message Decryption class